VSDPythonSDK/requirements.txt
cryptography==44.0.2
Pillow==11.1.0
Requests==2.32.3
websocket_client==1.8.0
